requirements.txt
anthropic
chromadb
streamlit
plotly
pandas
numpy
requests
beautifulsoup4
pypdf
sentence-transformers
faker
python-dotenv